requirements.txt
BitstampClient==2.2.8
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
urllib3==1.26.9
websocket-client==1.3.2
